double -> long double

+++ trunk/freqchar.h
@@ -1,9 +1,9 @@
 #ifndef FREQCHAR_H
 #define FREQCHAR_H
 
-void freqChar( double * T, double * F, int n, int m ) ;
-void magnitude( double * F, double * M, int n ) ;
-void phase( double * F, double * P, int n ) ;
+void freqChar( long double * T, long double * F, int n, int m ) ;
+void magnitude( long double * F, double * M, int n ) ;
+void phase( long double * F, double * P, int n ) ;
 void group( double * P, double * D, int n ) ;
 
 #endif // FREQCHAR_H
