+++ trunk/windowCheby.h
@@ -1,6 +1,6 @@
 #ifndef WINDOWCHEBY_H
 #define WINDOWCHEBY_H
 
-void chebyshev( double * W, int aLen, double aAtten ) ;
+void chebyshev( long double * W, int aLen, double aAtten ) ;
 
 #endif // WINDOWCHEBY_H
