+++ trunk/windowHarris.h
@@ -1,8 +1,8 @@
 #ifndef WINDOWHARRIS_H
 #define WINDOWHARRIS_H
 
-void harris( double * W, int aLen ) ;
-void nuttall( double * W, int aLen ) ;
-void flattop( double * W, int aLen ) ;
+void harris( long double * W, int aLen ) ;
+void nuttall( long double * W, int aLen ) ;
+void flattop( long double * W, int aLen ) ;
 
 #endif // WINDOWHARRIS_H
