+++ trunk/windowKaiser.h
@@ -1,6 +1,6 @@
 #ifndef WINDOWKAISER_H
 #define WINDOWKAISER_H
 
-void kaiser( double * W, int aLen, double aAtten ) ;
+void kaiser( long double * W, int aLen, double aAtten ) ;
 
 #endif // WINDOWKAISER_H
